+++ examples/07_multistore/src/Person.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 
-import { useReduxDispatch, useReduxState } from 'reactive-react-redux';
+import { useDispatch, useTrackedState } from 'reactive-react-redux';
 
 import { Action, State } from './state';
 
 const Person = () => {
-  const state = useReduxState<State>();
-  const dispatch = useReduxDispatch<Action>();
+  const state = useTrackedState<State>();
+  const dispatch = useDispatch<Action>();
   return (
     <div>
       <div>