+++ examples/09_thunk/src/Person.tsx
@@ -2,13 +2,13 @@ import * as React from 'react';
 import { Dispatch } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
 
-import { useReduxDispatch, useReduxState } from 'reactive-react-redux';
+import { useDispatch, useTrackedState } from 'reactive-react-redux';
 
 import { Action, State } from './state';
 
 const Counter: React.FC<{ firstName: string }> = ({ firstName }) => {
-  const state = useReduxState<State>();
-  const dispatch = useReduxDispatch<Action>();
+  const state = useTrackedState<State>();
+  const dispatch = useDispatch<Action>();
   return (
     <div>
       {Math.random()}
@@ -23,8 +23,8 @@ const Counter: React.FC<{ firstName: string }> = ({ firstName }) => {
 };
 
 const Person = () => {
-  const state = useReduxState<State>();
-  const dispatch = useReduxDispatch<Action>();
+  const state = useTrackedState<State>();
+  const dispatch = useDispatch<Action>();
   const setRandomFirstName = () => {
     const dispatchForThunk = dispatch as ThunkDispatch<State, {}, Action>;
     dispatchForThunk(async (d: Dispatch<Action>) => {