+++ examples/11_todolist/src/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { render } from 'react-dom';
-import { ReduxProvider } from 'reactive-react-redux';
+import { Provider } from 'reactive-react-redux';
 import { createStore } from 'redux';
 
 import rootReducer from './reducers';
@@ -9,8 +9,8 @@ import App from './components/App';
 const store = createStore(rootReducer);
 
 render(
-  <ReduxProvider store={store}>
+  <Provider store={store}>
     <App />
-  </ReduxProvider>,
+  </Provider>,
   document.getElementById('app'),
 );