+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Direct state context
-  - No breaking change in API, but it may behave differently
 - Support custom context
+- Rename hooks to be somewhat compatible with react-redux hooks
 
 ## [3.0.0] - 2019-05-18
 ### Changed